doctest fix

+++ src/lu.jl
@@ -15,10 +15,10 @@ julia> A = [1.2 2.3
 
 julia> ws = LUWs(A)
 LUWs
-  ipiv: 2-element Vector{Int64}
+  ipiv: 2-element Vector{Int}
 
 julia> t = LU(LAPACK.getrf!(ws, A)...)
-LU{Float64, Matrix{Float64}, Vector{Int64}}
+LU{Float64, Matrix{Float64}, Vector{Int}}
 L factor:
 2×2 Matrix{Float64}:
  1.0       0.0
